Validate null file, parse to float instead to int, update test, change to enum map

+++ src/main/web/js/controllers/statisticalDataController.js
@@ -55,7 +55,7 @@ app.controller("statisticalDataController", function($scope, $rootScope, request
     			return null;
     		};
 
-	        $scope.statisticalDataList[index].numberJsons.push({"number" : parseInt(number)});
+	        $scope.statisticalDataList[index].numberJsons.push({"number" : parseFloat(number)});
     	});
         return null;
     };
